kafka/requirements.txt
kafka-python
pymongo
dnspython 
python-dotenv